Reorganize bracket
